Changed air quality API

+++ app/src/main/java/bose/ankush/weatherify/domain/model/AirQuality.kt
@@ -2,7 +2,7 @@ package bose.ankush.weatherify.domain.model
 
 data class AirQuality(
     val aqi: Int?,
-    val co: Int?,
+    val co: Double?,
     val no2: Double?,
     val o3 : Double?,
     val so2: Double?,
